vuelta atras

+++ src/main/java/com/conexion/api/controller/ProductosController.java
@@ -40,9 +40,8 @@ public class ProductosController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Productos> obtenerPorId(@PathVariable Long id) {
+    public ResponseEntity<Productos> obtenerPorId(@PathVariable int id) {
         Optional<Productos> producto = productosRepository.findById(id);
         return producto.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
     }
-
 }
